+++ CHANGELOG.md
@@ -1,5 +1,30 @@
 # Changelog
 
+**8.1.0**
+
+```
+IMP: Integration tests
+IMP: Command 'post': New argument '--mode' to select the mode (odoo or raw, default is odoo)
+IMP: Command 'dblist' New argument '--only-active' to print only the active database name
+IMP: Screen: Pretty print array of objects and show the host in the prompt
+
+ADD: Unit tests
+ADD: RPC: Modified implementation to do rpc queries (issue #33)
+
+FIX: Command 'load': Return native promise
+FIX: Command 'context_term': Operation 'delete' and '--value' argument type changed to any
+FIX: Command 'context': '--value' argument type changed to any
+FIX: Command 'alias': Default operation
+FIX: Command 'call': Use correct properties (issue #34)
+FIX: Command 'longpolling': Don't print undefined value when do 'add_channel' operation
+FIX: Command 'uhg': Send group names as string
+FIX: Command 'lang': Moved to backend
+FIX: ParameterReader: Argument definition not found in 'validateAndFormat'
+FIX: ParameterReader: Correct detection of invalid arguments in subcommands
+FIX: Screen: Open view on click id (issue #32)
+FIX: CommandAssistant: Parameters with spaces and nav mode in 11.0
+```
+
 **8.0.0**
 
 ```